refactor: move model to domain
